reviews/src/main/java/com/udacity/course3/reviews/mongo/entity/MongoReview.java
package com.udacity.course3.reviews.mongo.entity;

import java.util.ArrayList;
import java.util.List;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

@Document(collection = "reviews")
public class MongoReview {
	@Id
	private String id;
	private String title;
	private String description;
	private String productId;
	private List<MongoComment> comments = new ArrayList<MongoComment>();

	public MongoReview() {
	}

	public MongoReview(String id, String title, String description, String productId) {
		this.id = id;
		this.title = title;
		this.description = description;
		this.productId = productId;
	}

	/**
	 * @return the id
	 */
	public String getId() {
		return id;
	}

	/**
	 * @param id the id to set
	 */
	public void setId(String id) {
		this.id = id;
	}

	/**
	 * @return the title
	 */
	public String getTitle() {
		return title;
	}

	/**
	 * @param title the title to set
	 */
	public void setTitle(String title) {
		this.title = title;
	}

	/**
	 * @return the description
	 */
	public String getDescription() {
		return description;
	}

	/**
	 * @param description the description to set
	 */
	public void setDescription(String description) {
		this.description = description;
	}

	/**
	 * @return the productId
	 */
	public String getProductId() {
		return productId;
	}

	/**
	 * @param productId the productId to set
	 */
	public void setProductId(String productId) {
		this.productId = productId;
	}

	/**
	 * @return the comments
	 */
	public List<MongoComment> getComments() {
		return comments;
	}

	/**
	 * @param comments the comments to set
	 */
	public void setComments(List<MongoComment> comments) {
		this.comments = comments;
	}
}